% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_unregistered_curves.R
\name{amp_curve}
\alias{amp_curve}
\title{Simulate amplitude variance}
\usage{
amp_curve(grid)
}
\arguments{
\item{grid}{Grid of x values over which to evaluate the function.}
}
\description{
This function generates amplitudes for simulated accelerometer data.
}
