% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfpca.R
\name{bfpca}
\alias{bfpca}
\title{Binary functional principal components analysis}
\usage{
bfpca(Y, npc = 1, Kt = 6, maxiter = 50, t_min = NULL, t_max = NULL,
  print.iter = FALSE, row_obj = NULL, seed = 1988, ...)
}
\arguments{
\item{Y}{Dataframe. Should have values id, value, index.}

\item{npc}{Defaults to 1. Number of principal components to calculate.}

\item{Kt}{Number of B-spline basis functions used to estimate mean functions. Defaults to 6.}

\item{maxiter}{Maximum number of iterations to perform for EM algorithm. Default is 20.}

\item{t_min}{Minimum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, Taken to be minimum observed value.}

\item{t_max}{Maximum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be maximum observed value.}

\item{print.iter}{Prints current error and iteration}

\item{row_obj}{If NULL, the function cleans the data and calculates row indices. Keep this NULL if you are using 
standalone \code{register} function.}

\item{seed}{Set seed for reproducibility. Seed value defaults to 1988.}

\item{...}{Additional arguments passed to or from other functions}
}
\value{
An object of class \code{fpca} containing:
\item{knots}{Cutpoints for B-spline basis used to rebuild \code{alpha}.}
\item{efunctions}{\eqn{D \times npc} matrix of estimated FPC basis functions.}
\item{evalues}{Estimated variance of the FPC scores.}
\item{npc}{number of FPCs.}
\item{scores}{\eqn{I\times npc} matrix of estimated FPC scores.}
\item{alpha}{Estimated population-level mean.}
\item{mu}{Estimated population-level mean. Same value as \code{alpha} but included for compatibility
with \code{refund.shiny} package.}
\item{subject_coefs}{B-spline basis coefficients used to construct subject-specific means. 
For use in \code{registr()} function.}
\item{Yhat}{FPC approximation of subject-specific means.}
\item{Y}{The observed data.}
\item{family}{\code{gaussian} or \code{binomial}, for compatibility with \code{refund.shiny} package.}
}
\description{
Function used in the FPCA step for registering binary functional data. 
This method uses a variational EM algorithm to estimate scores and principal components for 
binary functional data.
}
\examples{
\dontrun{
Y = simulate_functional_data()$Y
bfpca_object = bfpca(Y, npc = 2, print.iter = TRUE)
}

}
\references{
Jaakkola, T. S. and Jordan, M. I. (1997).
A variational approach to Bayesian logistic regression models and their extensions. 
\emph{A variational approach to Bayesian logistic regression models and their extensions.}.

Tipping, M. E. (1999). Probabilistic Visualisation of High-dimensional binary data.
\emph{Advances in neural information processing systems}, 592--598.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu},
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}
