% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blank_tibble}
\alias{blank_tibble}
\title{Construct a zero-row tibble with given column names}
\usage{
blank_tibble(col_names)
}
\arguments{
\item{col_names}{Character vector of column names}
}
\description{
Construct a zero-row tibble with given column names
}
\details{
All columns will be \code{logical} so that they don't disturb exiting
columns when appended with \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}}`
}
