% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinalise.R
\name{cardinalise}
\alias{cardinalise}
\title{Split field values of cardinality=n into vectors}
\usage{
cardinalise(field)
}
\arguments{
\item{field}{Character vector where each element contains zero or more values
separated by semicolons, e.g. \code{NA}, \code{""}, \code{"foo"}, \code{"foo;bar"},
\code{"foo;bar;baz"}, etc.}
}
\description{
Fields whose cardinality is \code{"n"} can hold multiple values, separated by
semicolons (\code{;}).  This function splits those into vectors, so the column
becomes a list-column where each cell is a vector.
}
\examples{
x <- c(NA, "", "foo", "foo;bar", "foo;bar;baz")
cardinalise(x)
tibble::tibble(x = cardinalise(x))
}
