% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraints}
\alias{constraints}
\title{Define constraints for optimization of warping functions}
\usage{
constraints(Kh)
}
\arguments{
\item{Kh}{number of B-spline basis functions used to estimate warping functions \emph{h}.}
}
\description{
Constraints ensure monotonicty of spline coefficients for warping functions for use with \code{constrOptim()} function.
}
