% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curie.R
\name{curie}
\alias{curie}
\alias{curie_fields}
\alias{curie_fields.register}
\alias{curie_links}
\alias{curie_links.default}
\alias{curie_links.register}
\title{List registers linked to by CURIEs}
\usage{
curie_fields(x)

\method{curie_fields}{register}(x)

curie_links(x)

\method{curie_links}{default}(x)

\method{curie_links}{register}(x)
}
\arguments{
\item{x}{Object of class \code{"register"}, or a character vector that is a field
of a register object (\code{register$data$foo}) of datatype \code{"curie"}.}
}
\value{
A character vector of names of registers that are linked to.
}
\description{
Values in fields of datatype \code{"curie"} link to other registers.  These
functions list which registers are linked to by a field or all the fields in
a register.

\code{\link[=curie_fields]{curie_fields()}} names which fields in a register are of datatype \code{"curie"}.

\code{\link[=curie_links]{curie_links()}} names which registers are linked to by a field of datatype
\code{"curie"}, or by a whole register that has fields of datatype \code{"curie"}.
}
\examples{
register <- rr_register("statistical-geography")
curie_fields(register)
curie_links(register$data$area)
curie_links(register$data$organisation)
curie_links(register)
}
