% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{Convert data to a \code{refund} object.}
\usage{
data_clean(data, family = "binomial")
}
\arguments{
\item{data}{Dataframe. Should have values id, value, index.}

\item{family}{\code{gaussian} or \code{binomial}.}
}
\description{
Function used for data cleaning.
}
