% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfpca_helpers.R
\name{deriv.inv.logit}
\alias{deriv.inv.logit}
\title{Estimate the derivative of the logit function}
\usage{
\method{deriv}{inv.logit}(x)
}
\arguments{
\item{x}{Value at which the derivative is computed}
}
\description{
Compute the derivative of the logit function for a given point \code{x}.
}
\author{
Alexander Bauer \email{alexander.bauer@stat.uni-muenchen.de}
}
