% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expectedScores}
\alias{expectedScores}
\title{Calculate expected score and score variance for the current subject.}
\usage{
expectedScores(Y, mu, psi, theta, theta_quad)
}
\arguments{
\item{Y}{vector of observations for the current subject.}

\item{mu}{vector of spline coefficients for the population mean.}

\item{psi}{matrix of spline coefficients for the principal component basis functions.}

\item{theta}{spline basis functions for the current subject.}

\item{theta_quad}{quadratic form of theta for the current subject.}
}
\value{
A list with expected score mean and variance for the current subject.
}
\description{
Calculations derived using maximum likelihood estimation.
}
