% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfpca_twoStep.R
\name{gfpca_twoStep}
\alias{gfpca_twoStep}
\title{Generalized functional principal component analysis}
\usage{
gfpca_twoStep(
  Y,
  family = "gaussian",
  npc = 1,
  Kt = 8,
  t_min = NULL,
  t_max = NULL,
  row_obj = NULL,
  index_relevantDigits = 2L,
  estimation_accuracy = "high",
  start_params = NULL,
  periodic = FALSE,
  ...
)
}
\arguments{
\item{Y}{Dataframe. Should have values id, value, index.}

\item{family}{One of \code{c("gaussian","binomial")}. Defaults to
\code{"gaussian"}.}

\item{npc}{Number of principal components to calculate. Defaults to 1.}

\item{Kt}{Number of B-spline basis functions used to estimate mean functions. Default is 8.}

\item{t_min}{Minimum value to be evaluated on the time domain.}

\item{t_max}{Maximum value to be evaluated on the time domain.}

\item{row_obj}{If NULL, the function cleans the data and calculates row indices. 
Keep this NULL if you are using standalone \code{register} function.}

\item{index_relevantDigits}{Positive integer \code{>= 2}, stating the number
of relevant digits to which the index grid should be rounded. Coarsening the
index grid is necessary since otherwise the covariance surface matrix
explodes in size in the presence of too many unique index values (which is
the case after some registration step). Defaults to 2. Set to \code{NULL} to
prevent rounding.}

\item{estimation_accuracy}{TODO}

\item{start_params}{Optional start values for gamm4.}

\item{periodic}{If TRUE, uses periodic b-spline basis functions. Default is FALSE.}

\item{...}{Additional arguments passed to registr and fpca functions.}
}
\value{
An object of class \code{fpca} containing:
\item{t_vec}{Time vector over which the mean \code{mu} was evaluated.}
\item{knots}{Cutpoints for B-spline basis used to rebuild \code{alpha}.}
\item{efunctions}{\eqn{D \times npc} matrix of estimated FPC basis functions.}
\item{evalues}{Estimated variance of the FPC scores.}
\item{evalues_sum}{Sum of all eigenvalues, before restricting the \code{evalues}
vector to the first \code{npc} elements.}
\item{npc}{number of FPCs.}
\item{scores}{\eqn{I \times npc} matrix of estimated FPC scores.}
\item{alpha}{Estimated population-level mean.}
\item{mu}{Estimated population-level mean. Same value as \code{alpha} but included for compatibility
with \code{refund.shiny} package.}
\item{subject_coefs}{Always \code{NA} but included for full consistency
with \code{fpca_gauss} and \code{bfpca}.} 
\item{Yhat}{FPC approximation of subject-specific means.}
\item{Y}{The observed data.}
\item{family}{\code{binomial}, for compatibility with \code{refund.shiny} package.}
\item{gamm4_theta}{Estimated parameters of the mixed model.}
}
\description{
Function for applying FPCA to different exponential family distributions,
used in the FPCA step for registering functional data,
called by \code{\link{register_fpca}} when \code{GFPCA_type = "two-step"}. \cr \cr
The method is partly based on the approach of Hall et al. (2008) to estimate
functional principal components. The full method is outlined as 
`two-step approach` in Gertheiss et al. (2017). \cr \cr
This function is an adaptation of the implementation of Jan
Gertheiss for Gertheiss et al. (2017), with focus on higher (RAM) efficiency
for large data settings.
}
\examples{
TODO

}
\references{
Gertheiss, J., Goldsmith, J., & Staicu, A. M. (2017). A note on
modeling sparse exponential-family functional response curves.
\emph{Computational statistics & data analysis}, 105, 46--52.

Hall, P., Müller, H. G., & Yao, F. (2008). Modelling sparse
generalized longitudinal observations with latent Gaussian processes.
\emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)},
70(4), 703--723.
}
\author{
Alexander Bauer \email{alexander.bauer@stat.uni-muenchen.de},
based on work of Jan Gertheiss
}
