% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambdaF}
\alias{lambdaF}
\title{Apply lambda transformation of variational parameter.}
\usage{
lambdaF(x)
}
\arguments{
\item{x}{The value to which you apply the function}
}
\value{
A numeric value that has been transformed.
}
\description{
Simple function for use within other C++ functions.
}
