% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_h.R
\name{loss_h}
\alias{loss_h}
\title{Loss function for registration step optimization}
\usage{
loss_h(Y, basis.tstar, mean.coefs, knots, beta.inner, family, t.min, t.max)
}
\arguments{
\item{Y}{vector of observed points.}

\item{basis.tstar}{B-spline basis for vector Y.}

\item{mean.coefs}{spline coefficient vector for mean curve.}

\item{knots}{knot locations for B-spline basis used to estimate mean and FPC basis function.}

\item{beta.inner}{spline coefficient vector to be estimated for warping function h.}

\item{family}{\code{gaussian} or \code{binomial}.}

\item{t.min}{minimum value to be evaluated on the time domain (useful if data are sparse and / or irregular).}

\item{t.max}{maximum value to be evaluated on the time domain (useful if data are sparse and / or irregular).}
}
\description{
Loss function for registration step optimization
}
