% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_h2.R
\name{loss_h2}
\alias{loss_h2}
\title{Loss function for registration step optimization}
\usage{
loss_h2(Y, tstar, mean_coefs, knots, family, t_min, t_max, alpha_beta)
}
\arguments{
\item{Y}{vector of observed points.}

\item{tstar}{observed time points}

\item{mean_coefs}{spline coefficient vector for mean curve.}

\item{knots}{knot locations for B-spline basis used to estimate mean and FPC basis function.}

\item{family}{\code{gaussian} or \code{binomial}.}

\item{t_min}{minimum value to be evaluated on the time domain.}

\item{t_max}{maximum value to be evaluated on the time domain.}

\item{alpha_beta}{alpha beta coefficients for warping function h.}
}
\description{
Loss function for registration step optimization
}
