% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_h_gradient.R
\name{loss_h_gradient}
\alias{loss_h_gradient}
\title{Gradient of loss function for registration step}
\usage{
loss_h_gradient(Y, Theta_phi, mean.coefs, knots, beta.inner,
  family = "gaussian", t_min, t_max)
}
\arguments{
\item{Y}{vector of observed points.}

\item{Theta_phi}{B-spline basis for vector Y.}

\item{mean.coefs}{spline coefficient vector for mean curve.}

\item{knots}{knot locations for B-spline basis used to estimate mean and FPC basis function.}

\item{beta.inner}{spline coefficient vector to be estimated for warping function h.}

\item{family}{\code{gaussian} or \code{binomial}.}

\item{t_min}{minimum value to be evaluated on the time domain (useful if data are sparse and / or irregular).}

\item{t_max}{maximum value to be evaluated on the time domain (useful if data are sparse and / or irregular).}
}
\description{
Gradient of loss function for registration step
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
