% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_apply_datatype.R
\name{maybe_parse_iso_8601}
\alias{maybe_parse_iso_8601}
\title{Parse ISO8601 strings as datetimes or characters}
\usage{
maybe_parse_iso_8601(x, parse = FALSE)
}
\arguments{
\item{x}{Character vector}

\item{parse}{Logicial, whether to parse as a datetime or leave as (or convert
to) character \code{\link[parsedate:parse_iso_8601]{parsedate::parse_iso_8601()}}}
}
\description{
Parse ISO8601 strings as datetimes or characters
}
