% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_unregistered_curves.R
\name{mean_curve}
\alias{mean_curve}
\title{Simulate mean curve}
\usage{
mean_curve(grid)
}
\arguments{
\item{grid}{Grid of x values over which to evaluate the function.}
}
\value{
A numeric vector.
}
\description{
This function generates mean for simulated accelerometer data.
}
