% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functional_data.R
\name{mean_sim}
\alias{mean_sim}
\title{Simulate mean}
\usage{
mean_sim(grid)
}
\arguments{
\item{grid}{Grid of x values over which to evaluate the function.}
}
\description{
This function generates mean for simulated functional data.
}
