% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing_col_names}
\alias{missing_col_names}
\title{Return column names missing from a data frame}
\usage{
missing_col_names(.data, ...)
}
\arguments{
\item{.data}{Data frame.}

\item{...}{names of columns.}
}
\value{
A character vector of column names that aren't in \code{.data}.
}
\description{
Return column names missing from a data frame
}
