% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_curie.R
\name{parse_curie}
\alias{parse_curie}
\title{Parse a CURIE into prefix and reference}
\usage{
parse_curie(curie)
}
\arguments{
\item{curie}{Character of the form \code{"prefix:reference"} or \code{"prefix:"}.}
}
\description{
Parse a CURIE into prefix and reference
}
\examples{
parse_curie("country:GB")
parse_curie("country:")
parse_curie("country")
}
