% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_linear2_hinv.R
\name{piecewise_linear2_hinv}
\alias{piecewise_linear2_hinv}
\title{Create two-parameter piecewise linear (inverse) warping functions}
\usage{
piecewise_linear2_hinv(
  grid,
  knot1_x = 0.25,
  knot1_y = 0.3,
  knot2_x = 0.75,
  knot2_y = 0.9
)
}
\arguments{
\item{grid}{grid of values over which to evaluate the function.}

\item{knot1_x}{controls the x-location of the first knot. Defaults is \code{0.25}.}

\item{knot1_y}{controls the y-location of the first knot. Defaults is \code{0.3}.}

\item{knot2_x}{controls the x-location of the second knot. Defaults is \code{0.75}.}

\item{knot2_y}{controls the y-location of the second knot. Defaults is \code{0.9}.}
}
\description{
This function uses a 2-knot piecewise linear model to calculate inverse warping 
functions for registration. The parameters \code{knot1_x} and \code{knot1_y}
control the x and y locations of the first knot, and the parameters
\code{knot1_x} and \code{knot1_y} control the x and y locations of the second
knot. The designation (inverse) is intended to communicate that these 
functions take data from the unregistered space to the registered space, 
consistent with #' functional data literature on registration.
}
\author{
Erin McDonnell \email{eim2117@cumc.columbia.edu}

Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
