% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_linear_hinv.R
\name{piecewise_linear_hinv}
\alias{piecewise_linear_hinv}
\title{Create two-parameter piecewise linear (inverse) warping functions}
\usage{
piecewise_linear_hinv(grid, knot_x = 0.5, knot_y = 0.5)
}
\arguments{
\item{grid}{grid of values over which to evaluate the function.}

\item{beta1}{parameter that controls the slope from time 0 to the knot. 
Result approaches identity warp as beta approaches 1.}

\item{knot}{controls where the result crosses the identity warp. 
Default is 0.5.}
}
\description{
This function uses a piecewise linear model to calculate inverse warping 
functions for #' registration. The parameter \code{knot_x} controls the 
x-location of the knot, #' and the parameter \code{knot_y} the y-location 
of the knot. The designation (inverse) is intended to communicate that these 
functions take data from the unregistered space to the registered space, 
consistent with #' functional data literature on registration.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
