% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binary.R
\name{plot_binary}
\alias{plot_binary}
\title{Visualize results of binary registration}
\usage{
plot_binary(reg_obj, alpha = 0.25, tstar, t.min = NULL, t.max = NULL)
}
\arguments{
\item{reg_obj}{registration object to be plotted.}

\item{alpha}{alpha value for ggplot.}

\item{tstar}{vector of original observed time values.}

\item{t.min}{minimum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be minimum observed value.}

\item{t.max}{maximum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be maximum observed value.}
}
\description{
Function takes in registration object and plots unregistered kernel smoothed curves, registered kernel smoothed curves, 
and warping functions.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
