% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpca.R
\name{plot_fpca}
\alias{plot_fpca}
\title{Plot the results of a functional PCA}
\usage{
plot_fpca(
  fpca_obj,
  plot_npc = fpca_obj$npc,
  var_factor = 2,
  response_function = function(x) {     x },
  subtitle = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = "t [registered]",
  ylab = "y"
)
}
\arguments{
\item{fpca_obj}{Object of class \code{"fpca"}.}

\item{plot_npc}{Optional number of the main FPCs to be plotted. Defaults to
all FPCs contained in \code{fpca_obj}.}

\item{var_factor}{Numeric factor with which the FPC's are multiplied
to display their variation in the plots. Defaults to 2, but can be set to
2 times the standard deviation of the obtained FPC scores.}

\item{response_function}{Optional response function to be applied before
plotting the curves. Defaults to the identity function.}

\item{subtitle}{If TRUE (default) the parameter \code{var_factor}
is displayed in the plot subtitle.}

\item{xlim, ylim}{Optional numeric vectors with limits for the x and y axis.}

\item{xlab, ylab}{Optional titles for the x and y axis.}
}
\value{
Grid of \code{ggplot} plots, created with \code{cowplot::plot_grid}.
}
\description{
Plot FPCA results by visualizing the variation of the individual FPCs around
the global mean. based on an object created with function
\code{\link{fpca_gauss}}, \code{\link{bfpca}} or \code{\link{gfpca_twoStep}}. \cr \cr
The shares of explained variance are included in the plot titles if
\code{fpca_obj} contains an element \code{evalues_sum}. This is currently
only the case for FPCA objects created with \code{\link{gfpca_twoStep}}.
}
\examples{
data(growth_incomplete)

fpca_obj = fpca_gauss(Y = growth_incomplete, npc = 2)
plot_fpca(fpca_obj)

}
\author{
Alexander Bauer \email{alexander.bauer@stat.uni-muenchen.de}
}
