% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_registration.R
\name{plot_registration}
\alias{plot_registration}
\title{Visualize results of registration}
\usage{
plot_registration(reg_obj, fpca_obj = NULL, family = "gaussian",
  alpha = 0.25, tstar, t.min = NULL, t.max = NULL, Kt = NULL)
}
\arguments{
\item{reg_obj}{registration object to be plotted.}

\item{fpca_obj}{fpca object that contains mean values (fitted values).}

\item{family}{\code{gaussian} or \code{binomial}.}

\item{alpha}{alpha value for ggplot.}

\item{tstar}{vector of original observed time values.}

\item{t.min}{minimum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be minimum observed value.}

\item{t.max}{maximum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be maximum observed value.}

\item{Kt}{number of B-spline basis functions used to estimate coefficient functions. Defaults to 10.}
}
\description{
Function takes in registration object and plots unregistered curves, registered curves, 
and warping functions. Note that there are a lot of specifications right now that only make sense with 
a simulation.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
