% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functional_data.R
\name{psi1_sim}
\alias{psi1_sim}
\title{Simulate PC1}
\usage{
psi1_sim(grid)
}
\arguments{
\item{grid}{Grid of x values over which to evaluate the function.}
}
\description{
This function generates the first principal component for simulated functional data.
}
