% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functional_data.R
\name{psi2_sim}
\alias{psi2_sim}
\title{Simulate PC2}
\usage{
psi2_sim(grid)
}
\arguments{
\item{grid}{Grid of x values over which to evaluate the function.}
}
\description{
This function generates the second principal component for simulated functional data.
}
