% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_basic.R
\name{reg_basic}
\alias{reg_basic}
\title{Register curves from exponential family}
\usage{
reg_basic(data = NULL, mean.coefs = NULL, Kt = 10, Kh = 5,
  family = "gaussian", gradient = TRUE, t.min = NULL, t.max = NULL)
}
\arguments{
\item{data}{output of data_clean function}

\item{mean.coefs}{If null, calculates the population mean from the data.}

\item{Kt}{number of B-spline basis functions used to estimate mean functions. Defaults to 10.}

\item{Kh}{number of B-spline basis functions used to estimate warping functions \emph{h}. Defaults to 5.}

\item{family}{\code{gaussian} or \code{binomial}.}

\item{gradient}{if TRUE, uses analytic gradient to calculate derivative. 
If FALSE, calculates gradient nuemrically.}

\item{t.min}{minimum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be minimum observed value.}

\item{t.max}{maximum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be maximum observed value.}
}
\description{
Function used in the registration step. Assumes you feed it an object from data_clean, mean coefficients.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
