% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{Register curves from exponential family}
\usage{
register(obj = NULL, Y = NULL, Kt = 10, Kh = 5, family = "gaussian",
  gradient = TRUE, t.min = NULL, t.max = NULL, row_obj = NULL)
}
\arguments{
\item{obj}{current estimate of FPC objects (including spline coefs and scores). Can be NULL only if Y argument is selected.}

\item{Y}{data matrix, each row is a subject.}

\item{Kt}{number of B-spline basis functions used to estimate mean functions. Defaults to 10.}

\item{Kh}{number of B-spline basis functions used to estimate warping functions \emph{h}. Defaults to 5.}

\item{family}{\code{gaussian} or \code{binomial}.}

\item{gradient}{if TRUE, uses analytic gradient to calculate derivative. 
If FALSE, calculates gradient nuemrically.}

\item{t.min}{minimum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be minimum observed value.}

\item{t.max}{maximum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be maximum observed value.}

\item{row_obj}{if NULL, the function cleans the data and calculates row indices. Keep this NULL if you are using 
standalone \code{register} function.}
}
\description{
Function used in the registration step of an FPCA-based approach for registering exponential-family functional data. 
This method uses constrained optimization to estimate spline coefficients for warping functions,
where the objective function for optimization comes from maximizing the EF likelihood 
subject to monotonicity constraints on the warping functions. You have to either specify \code{obj}, which is a fpca 
object from an earlier step, or \code{Y}, which can be a matrix in wide format or a dataframe in long format. If 
\code{Y} is a dataframe, the subjects IDs, times, and observations should have column names id, index, and value, 
respectively. If Y is input, then we automatically calculate a template using glm.
}
\examples{

\dontrun{
   register_step = register(obj = NULL, Y = Y_sim, Kt = 10, Kh = 5, family = "binomial", gradient = TRUE)
}

}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
