% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_fpca.R
\name{register_fpca}
\alias{register_fpca}
\title{Register curves from exponential family using constrained optimization and generalized FPCA}
\usage{
register_fpca(Y, Kt = 10, Kh = 4, family = "binomial",
  max_iterations = 20, npc = 1, ...)
}
\arguments{
\item{Y}{dataframe. Should have values id, value, index.}

\item{Kt}{number of B-spline basis functions used to estimate mean functions. 
Defaults to 10.}

\item{Kh}{number of B-spline basis functions used to estimate warping functions \emph{h}. Defaults to 5.}

\item{family}{\code{gaussian} or \code{binomial}.}

\item{max_iterations}{number of iterations between fpca step and registration step.}

\item{npc}{defaults to 1. Number of principal components to calculate.}

\item{...}{additional arguments passed to registr and fpca functions}
}
\value{
An object of class \code{registration} containing:
\item{fpca_obj}{List of items from FPCA step.}
\item{Y}{The observed data plus variables \code{t_star} and \code{t_hat} which are the
unregistered grid and registered grid, respectively.}
\item{time_warps}{List of time values for each iteration of the algorithm. 
\code{time_warps[1]} is the original (observed) time
and \code{time_warps[n]} provides time values for the nth iteration.}
\item{loss}{Loss for each iteration of the algorithm, calculated in the registration step using an 
exponential family likelihood with natural parameter from the FPCA step.}

family \code{gaussian} or \code{binomial}.
}
\description{
Function combines constrained optimization and FPCA to estimate warping functions for 
exponential family curves.
}
\examples{

\dontrun{
 Y = simulate_unregistered_curves(I = 50, D = 200)
 registr_object = register_fpca(Y, family = "binomial", max_iterations = 25)
}

}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}
