% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{register_lines}
\alias{register_lines}
\title{Download or open a register file}
\usage{
register_lines(file, phase = c("beta", "alpha"), path_type = c("url",
  "file"))
}
\arguments{
\item{file}{Character, either the name of the register, or a file path.}

\item{phase}{Character, one of \code{"beta"} or \code{"alpha"}}

\item{path_type}{Character, one of \code{"url"} or \code{"file"} to decide what to do
with \code{register}.}
}
\description{
Download or open a register file
}
\details{
Either download a register, constructing the URL from the name of
the register and the given phase, or load it from a file containing the RSF
(register serialisation format).
}
