% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R, R/registr-package.R
\docType{package}
\name{registr-package}
\alias{registr-package}
\title{Register Exponential Family Functional Data}
\description{
Software for registering functional data from the exponential family of distributions.

A method for performing joint registration and functional principal component analysis for curves (functional data) that are generated from exponential family distributions. This mainly implements the algorithms described in 'Wrobel et al. (2019)' \doi{10.1111/biom.12963} and further adapts them to potentially incomplete curves where (some) curves are not observed from the beginning and/or until the end of the common domain. Curve registration can be used to better understand patterns in functional data by separating curves into phase and amplitude variability. This software handles both binary and continuous functional data, and is especially applicable in accelerometry and wearable technology.
}
\author{
Julia Wrobel

\strong{Maintainer}: Julia Wrobel \email{julia.wrobel@emory.edu} (\href{https://orcid.org/0000-0001-6783-1421}{ORCID})

Authors:
\itemize{
  \item Alexander Bauer \email{alexander.bauer@stat.uni-muenchen.de}
  \item Erin McDonnell \email{eim2117@cumc.columbia.edu}
  \item Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}

Other contributors:
\itemize{
  \item Fabian Scheipl \email{fabian.scheipl@lmu.de} [contributor]
}

}
\keyword{internal}
