% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registr2.R
\name{registr2}
\alias{registr2}
\title{Register curves from exponential family}
\usage{
registr2(obj = NULL, Y = NULL, Kt = 8, family = "binomial",
  gradient = TRUE, t_min = NULL, t_max = NULL, row_obj = NULL, ...)
}
\description{
Register curves from exponential family
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
