% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registr.R
\name{registr_oneCurve}
\alias{registr_oneCurve}
\title{Internal function to register one curve}
\usage{
registr_oneCurve(i, arg_list, ...)
}
\arguments{
\item{i}{Numeric index of the curve under focus.}

\item{arg_list}{Named list of all arguments necessary for the registration
step.}

\item{...}{additional arguments passed to or from other functions}
}
\value{
An list containing:
\item{beta_new}{Estimated parameter vector of the warping function.}
\item{t_hat}{Vector of registered time domain.}
\item{loss}{Loss of the optimal solution.}
}
\description{
This internal function is only to be used from within \code{registr}.
It performs the main optimization step with \code{constrOptim} for the
registration of one curve.
}
\author{
Alexander Bauer \email{alexander.bauer@stat.uni-muenchen.de}
}
