% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{registration.fpca}
\alias{registration.fpca}
\alias{registration.fpca-package}
\title{registration.fpca- a package for exponential family registration that I am working on.}
\description{
It does registration stuff
}
\author{
Julia Wrobel
}
