% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_curie.R
\name{rr_curie}
\alias{rr_curie}
\alias{rr_curie_fields}
\alias{rr_curie_fields.register}
\alias{rr_curie_links}
\alias{rr_curie_links.default}
\alias{rr_curie_links.register}
\title{List registers linked to by CURIEs}
\usage{
rr_curie_fields(x)

\method{rr_curie_fields}{register}(x)

rr_curie_links(x)

\method{rr_curie_links}{default}(x)

\method{rr_curie_links}{register}(x)
}
\arguments{
\item{x}{Object of class \code{"register"}, or a character vector that is a field
of a register object (\code{register$data$foo}) of datatype \code{"curie"}.}
}
\value{
A character vector of names of registers that are linked to.
}
\description{
Values in fields of datatype \code{"curie"} link to other registers.  These
functions list which registers are linked to by a field or all the fields in
a register.

\code{\link[=rr_curie_fields]{rr_curie_fields()}} names which fields in a register are of datatype
\code{"curie"}.

\code{\link[=rr_curie_links]{rr_curie_links()}} names which registers are linked to by a field of datatype
\code{"curie"}, or by a whole register that has fields of datatype \code{"curie"}.
}
\examples{
register <- rr_register("statistical-geography")
rr_curie_fields(register)
rr_curie_links(register$data$area)
rr_curie_links(register$data$organisation)
rr_curie_links(register)
}
