% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_index.R
\name{rr_index}
\alias{rr_index}
\title{Index a register by CURIEs}
\usage{
rr_index(register, key = NULL)
}
\arguments{
\item{register}{An object of class \code{"register"}.}

\item{key}{Character, the name of the field to use as theh \code{reference} part
of the CURIE.  If ommitted (default) then the 'key' field of the register is
used (the field with the same name as the register, that is guaranteed to
have unique values)}
}
\description{
You can refer to records in a register by CURIEs of the form
\code{"prefix:reference"}, where the \code{prefix} is the name of the register, and
the \code{reference} is a value in a field of the register.  \code{\link[=rr_index]{rr_index()}}
indexes records by these CURIES.
}
\examples{
country <- rr_register("country")
rr_index(country, "start-date")
rr_index(country, "end-date")
rr_index(country)
rr_index(sg, "area")
}
