% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_key.R
\name{rr_key}
\alias{rr_key}
\alias{rr_key_fields}
\alias{rr_key_fields.register}
\alias{rr_key_links}
\alias{rr_key_links.register}
\title{List registers linked to by key fields}
\usage{
rr_key_fields(x)

\method{rr_key_fields}{register}(x)

rr_key_links(x)

\method{rr_key_links}{register}(x)
}
\arguments{
\item{x}{Object of class \code{"register"}, or a character vector that is a field
of a register object (\code{register$data$foo}) with the \code{"register"} property.}
}
\value{
A character vector of names of registers that are linked to.
}
\description{
Values in fields with the \code{"register"} proporty link to other registers.
These functions list which registers are linked to by a field or all the
fields in a register.

\code{\link[=rr_key_fields]{rr_key_fields()}} names which fields in a register have the \code{"register"}
properety.

\code{\link[=rr_key_links]{rr_key_links()}} names all fields in a register that have the \code{"register"}
property, and the name of the register each one links to.
}
\examples{
register <- rr_register("allergen")
rr_key_fields(register)
rr_key_links(register)
}
