% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_links.R
\name{rr_links}
\alias{rr_links}
\title{Links between registers}
\usage{
rr_links(x)
}
\arguments{
\item{x}{Object of class \code{"register"}, or a list of objects of class
\code{"register"}.  If a list of registers, the graph of all links between all
registers will be returned.}
}
\description{
Registers can link to each other either by a field with the \code{"register"}
property, or via CURIEs.  See \code{\link[=rr_key_links]{rr_key_links()}} and \code{\link[=rr_curie_links]{rr_curie_links()}} for
more about that.  \code{\link[=rr_links]{rr_links()}} returns a data frame of edges.
}
\examples{
sg <- rr_register("statistical-geography")
rr_links(sg)

allergen <- rr_register("allergen")
rr_links(allergen)

registers <- rr_registers()
rr_links(registers)

if (require(tidygraph)) {
  as_tbl_graph(rr_links(rr_register("statistical-geography"))) \%>\%
    ggraph(layout = "nicely") +
      geom_edge_fan(aes(alpha = ..index..), show.legend = FALSE) +
      geom_edge_loop() +
      geom_node_label(aes(label = name))

  as_tbl_graph(rr_links(rr_registers())) \%>\%
    ggraph(layout = "nicely") +
      geom_edge_fan(aes(alpha = ..index..), show.legend = FALSE) +
      geom_edge_loop() +
      geom_node_label(aes(label = name))
}
}
