% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_parse_curie.R
\name{rr_parse_curie}
\alias{rr_parse_curie}
\title{Parse a CURIE into prefix and reference}
\usage{
rr_parse_curie(curie)
}
\arguments{
\item{curie}{Character of the form \code{"prefix:reference"} or \code{"prefix:"}.}
}
\description{
Parse a CURIE into prefix and reference
}
\examples{
rr_parse_curie("country:GB")
rr_parse_curie("country:")
rr_parse_curie("country")
}
