% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_records.R
\name{rr_records}
\alias{rr_records}
\title{Download All Register Records}
\usage{
rr_records(register, phase = "beta", page_size = 5000)
}
\arguments{
\item{register}{character, name of the register, e.g. "school-eng"}

\item{phase}{character, one of "beta", "alpha", "discovery", default: "beta"}

\item{page_size}{numeric, number of records per page of download, default: 5000}
}
\value{
A tibble of all the records in a register
}
\description{
Download all records from a register, automatically handling
             pages of up to 5000 records each.
}
\details{
Does *not* include all entries, only the latest entry per record.
}
\examples{
\dontrun{
if(interactive()){
 get_all_records("country", "beta")
 }
}
}
