% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_resolve.R
\name{rr_resolve}
\alias{rr_resolve}
\alias{rr_resolve_key_field}
\alias{rr_resolve_curie_field}
\alias{rr_resolve_links}
\alias{rr_resolve_links.register}
\title{Resolve links in registers}
\usage{
rr_resolve_key_field(field_name, register, registers)

rr_resolve_curie_field(field_name, register, registers = NULL,
  targets = NULL)

rr_resolve_links(register, registers, targets = NULL)

\method{rr_resolve_links}{register}(register, registers, targets = NULL)
}
\arguments{
\item{field_name}{Character, the name of a field to resolve.}

\item{register}{An object of class \code{"register"}.}

\item{registers}{A list of objects of class \code{"register"}.  Links will be
sought among these registers.}

\item{targets}{An optional data frame of indexed register records.  This is
constructed internally to resolve CURIE fields, but is expensive, so you can
create your own as follows, and use it more than once.

targets <-
registers %>%
purrr::map(rr_snapshot) %>%
purrr::map_dfr(rr_index)}
}
\description{
Registers can link to each other either by a field with the \code{"register"}
property, or via CURIEs.  See \code{\link[=rr_key_links]{rr_key_links()}} and \code{\link[=rr_curie_links]{rr_curie_links()}} for
more about that.  These functions resolve those links.

When a link is resolve, the whole record is returned in a data frame.  That's
because the link doesn't specify a particular field of a record.  For fields
with \code{cardinality = "n"}, several records might be returned.  If the field is
also \code{datatype = "curie"}, then a list of single-row data frames will be
returned, because each record could be from a different register with a
different set of fields.

Resolve a key field

Resolve a CURIE field

Resolve all links in a register, both key and CURIE
}
\examples{
registers <- rr_registers()

targets <-
  registers \%>\%
  purrr::map(rr_snapshot) \%>\%
  purrr::map_dfr(rr_index)

sg <- registers$`statistical-geography`
rr_links(sg)

rr_resolve_key_field("register", sg, registers)$data
rr_resolve_curie_field("organisation", sg, registers)$data
rr_resolve_curie_field("organisation", sg, targets = targets)$data
rr_resolve_links(sg, registers)$data
rr_resolve_links(sg, registers, targets = targets)$data
}
