% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_rows.R
\name{rr_rows}
\alias{rr_rows}
\title{Get the number of records in a register}
\usage{
rr_rows(register, phase)
}
\arguments{
\item{register}{character, name of the register, e.g. "school-eng"}

\item{phase}{character, one of "beta", "alpha", "discovery", default: "beta"}
}
\value{
a numeric vector of length 1
}
\description{
Queries the API for the number of records in a register, without
    having to download the records themselves.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
