% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_rsf.R
\name{rr_rsf}
\alias{rr_rsf}
\title{Download a register RSF file}
\usage{
rr_rsf(name = NULL, phase = c("beta", "alpha"), file = NULL,
  write = FALSE, dest_path = NULL, quiet = FALSE)
}
\arguments{
\item{name}{Character, the name of the register.}

\item{phase}{Character, one of \code{"beta"} or \code{"alpha"}}

\item{file}{Character, file path or URL, passed on to \code{\link[readr:read_lines]{readr::read_lines()}}
if \code{name} is not provided.}

\item{write}{Logical, whether to write the RSF file to disk.  If \code{TRUE},
either \code{name} or \code{dest_path} must be provided.}

\item{dest_path}{Character, path and file name to write the RSF to.}

\item{quiet}{Logical, if \code{TRUE} does not print messages to the console.}
}
\description{
Either downloads the Register Serialisation Format (RSF) file of a register,
constructing the URL from the name of the register and the given phase, or
reads it from disk.  Optionally saves it to disk.
}
\examples{
rr_rsf("country")
}
