% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squareTheta}
\alias{squareTheta}
\title{Calculate quadratic form of spline basis functions for the current subject.}
\usage{
squareTheta(xi, theta)
}
\arguments{
\item{xi}{vector of variational parameters for the current subject.}

\item{theta}{spline basis functions for the current subject.}
}
\value{
A matrix of the quadratic form of theta for the current subject.
}
\description{
Calculations quadratic form of theta with diagonalized variational parameter in the center.
}
