% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_datatype.R
\name{apply_datatype}
\alias{apply_datatype}
\title{Apply registers-style datatypes to fields}
\usage{
apply_datatype(x, datatype)
}
\arguments{
\item{x}{Character vector (a field of a register).}

\item{datatype}{Name of the datatype to apply: currently one of \code{"curie"}, \code{"url"},
\code{"datetime"}, \code{"string"} and \code{"text"}.}
}
\description{
Apply registers-style datatypes to fields
}
\details{
Registers use their own datatypes. This function converts strings to
the equivalent R datatypes.
}
\examples{
apply_datatype("2014-04", "datetime")
}
