% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfpca.R
\name{bfpca}
\alias{bfpca}
\title{Binary functional principal components analysis}
\usage{
bfpca(Y, index = NULL, id = NULL, npc = 1, Kt = 6, maxiter = 20,
  t_min = NULL, t_max = NULL, print.iter = FALSE, row_obj = NULL,
  seed = 1988, ...)
}
\arguments{
\item{Y}{Input data.}

\item{index}{Defaults to NULL. Indicates which column is the x-axis if Y input is a dataframe.}

\item{id}{Defaults to NULL. Indicates which column gives the subject ids if Y input is a dataframe.}

\item{npc}{Defaults to 1. Number of principal components to calculate.}

\item{Kt}{Number of B-spline basis functions used to estimate mean functions. Defaults to 6.}

\item{maxiter}{Maximum number of iterations to perform for EM algorithm. Default is 20.}

\item{t_min}{Minimum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, Taken to be minimum observed value.}

\item{t_max}{Maximum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be maximum observed value.}

\item{print.iter}{Prints current error and iteration}

\item{row_obj}{If NULL, the function cleans the data and calculates row indices. Keep this NULL if you are using 
standalone \code{register} function.}

\item{seed}{Set seed for reproducibility. Seed value defaults to 1988.}

\item{...}{Additional arguments passed to or from other functions}
}
\description{
Function used in the FPCA step for registering binary functional data. 
This method uses a variational EM algorithm to estimate scores and principal components for 
binary functional data. The returned values subject_coef are subject specific means.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
