% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfpca.R
\name{bfpca}
\alias{bfpca}
\title{Binary functional principal components analysis}
\usage{
bfpca(Y, index = NULL, id = NULL, npc = 1, Kt = 10, maxiter = 30,
  t.min = NULL, t.max = NULL, print.iter = FALSE, row_obj = NULL)
}
\arguments{
\item{Y}{input data. Can be long format dataframe or wide format matrix. If you provide a matrix}

\item{index}{defaults to NULL. Indicates which column is the x-axis if Y input is a dataframe.}

\item{id}{defaults to NULL. Indicates which column gives the subject ids if Y input is a dataframe.}

\item{npc}{defaults to 1. Number of principal components to calculate.}

\item{Kt}{number of B-spline basis functions used to estimate coefficient functions. Defaults to 10.}

\item{maxiter}{maximum number of iterations to perform.}

\item{t.min}{minimum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be minimum observed value.}

\item{t.max}{maximum value to be evaluated on the time domain (useful if data are sparse and / or irregular). 
if `NULL`, taken to be maximum observed value.}

\item{print.iter}{prints current error and iteration}

\item{row_obj}{if NULL, the function cleans the data and calculates row indices. Keep this NULL if you are using 
standalone \code{register} function.}
}
\description{
Function used in the FPCA step for registering binary functional data. 
This method uses a variational EM algorithm to estimates scores and principal components for 
binary functional data. The returned values subject_coef are subject specific means.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
