% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{Convert data to a \code{refund} object}
\usage{
data_clean(data)
}
\arguments{
\item{data}{Dataframe. Should have values id, value, index.}
}
\value{
An list containing:
\item{Y}{The original data sorted by id and index.}
\item{Y_rows}{A dataframe containing the first and last row for each subject.}
}
\description{
Function used for data cleaning.
}
