% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfpca_twoStep.R
\name{gfpca_twoStep}
\alias{gfpca_twoStep}
\title{Generalized functional principal component analysis}
\usage{
gfpca_twoStep(
  Y,
  family = "gaussian",
  npc = 1,
  Kt = 8,
  t_min = NULL,
  t_max = NULL,
  row_obj = NULL,
  index_relevantDigits = 4L,
  estimation_accuracy = "high",
  start_params = NULL,
  periodic = FALSE,
  ...
)
}
\arguments{
\item{Y}{Dataframe. Should have values id, value, index.}

\item{family}{One of \code{c("gaussian","binomial","gamma")}. Defaults to
\code{"gaussian"}.}

\item{npc}{Number of principal components to calculate. Defaults to 1.}

\item{Kt}{Number of B-spline basis functions used to estimate mean functions. Default is 8.}

\item{t_min}{Minimum value to be evaluated on the time domain.}

\item{t_max}{Maximum value to be evaluated on the time domain.}

\item{row_obj}{If NULL, the function cleans the data and calculates row indices. 
Keep this NULL if you are using standalone \code{register} function.}

\item{index_relevantDigits}{Positive integer \code{>= 2}, stating the number
of relevant digits to which the index grid should be rounded. Coarsening the
index grid is necessary since otherwise the covariance surface matrix
explodes in size in the presence of too many unique index values (which is
always the case after some registration step). Defaults to 4. Set to
\code{NULL} to prevent rounding.}

\item{estimation_accuracy}{One of \code{c("high","low")}. When set to \code{"low"},
the mixed model estimation step in \code{lme4} is performed with lower
accuracy, reducing computation time. Defaults to \code{"high"}.}

\item{start_params}{Optional start values for gamm4.}

\item{periodic}{Only contained for full consistency with \code{fpca_gauss}
and \code{bfpca}. If TRUE, returns the knots vector for periodic b-spline
basis functions. Defaults to FALSE. This parameter does not change the
results of the two-step GFPCA.}

\item{...}{Additional arguments passed to \code{\link{cov_hall}}.}
}
\value{
An object of class \code{fpca} containing:
\item{t_vec}{Time vector over which the mean \code{mu} was evaluated.
The resolution is can be specified by setting \code{index_relevantDigits}.}
\item{knots}{Cutpoints for B-spline basis used to rebuild \code{alpha}.}
\item{efunctions}{\eqn{D \times npc} matrix of estimated FPC basis functions.}
\item{evalues}{Estimated variance of the FPC scores.}
\item{evalues_sum}{Sum of all eigenvalues, before restricting the \code{evalues}
vector to the first \code{npc} elements.}
\item{npc}{number of FPCs.}
\item{scores}{\eqn{I \times npc} matrix of estimated FPC scores.}
\item{alpha}{Estimated population-level mean.}
\item{mu}{Estimated population-level mean. Same value as \code{alpha} but included for compatibility
with \code{refund.shiny} package.}
\item{subject_coefs}{Always \code{NA} but included for full consistency
with \code{fpca_gauss} and \code{bfpca}.} 
\item{Yhat}{FPC approximation of subject-specific means.}
\item{Y}{The observed data.}
\item{family}{\code{binomial}, for compatibility with \code{refund.shiny} package.}
\item{gamm4_theta}{Estimated parameters of the mixed model.}
}
\description{
Function for applying FPCA to different exponential family distributions.
Used in the FPCA step for registering functional data,
called by \code{\link{register_fpca}} when \code{GFPCA_type = "two-step"}. \cr \cr
The method implements the `two-step approach` of Gertheiss et al. (2017)
and is based on the approach of Hall et al. (2008) to estimate functional
principal components. \cr \cr
This function is an adaptation of the implementation of Jan
Gertheiss for Gertheiss et al. (2017), with focus on higher (RAM) efficiency
for large data settings.
}
\examples{
data(growth_incomplete)

fpca_obj = gfpca_twoStep(Y = growth_incomplete, npc = 2, family = "gaussian")
plot_fpca(fpca_obj)

}
\references{
Gertheiss, J., Goldsmith, J., & Staicu, A. M. (2017). A note on
modeling sparse exponential-family functional response curves.
\emph{Computational statistics & data analysis}, 105, 46--52.

Hall, P., Müller, H. G., & Yao, F. (2008). Modelling sparse
generalized longitudinal observations with latent Gaussian processes.
\emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)},
70(4), 703--723.
}
\author{
Alexander Bauer \email{alexander.bauer@stat.uni-muenchen.de},
based on work of Jan Gertheiss
}
