% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_h.R
\name{loss_h}
\alias{loss_h}
\title{Loss function for registration step optimization}
\usage{
loss_h(
  Y,
  Theta_h,
  mean_coefs,
  knots,
  beta.inner,
  family,
  t_min,
  t_max,
  periodic = FALSE,
  Kt = 8,
  warping = "nonparametric",
  prior_1_x_mean_sd = NULL,
  prior_1_y_mean_sd = NULL,
  prior_2_x_mean_sd = NULL,
  prior_2_y_mean_sd = NULL
)
}
\arguments{
\item{Y}{vector of observed points.}

\item{Theta_h}{B-spline basis for inverse warping functions.}

\item{mean_coefs}{spline coefficient vector for mean curve.}

\item{knots}{knot locations for B-spline basis used to estimate mean and FPC basis function.}

\item{beta.inner}{spline coefficient vector to be estimated for warping function h.}

\item{family}{\code{gaussian} or \code{binomial}.}

\item{t_min}{minimum value to be evaluated on the time domain.}

\item{t_max}{maximum value to be evaluated on the time domain.}

\item{periodic}{If \code{TRUE} uses periodic b-spline basis functions. Default is \code{FALSE}.}

\item{Kt}{Number of B-spline basis functions used to estimate mean functions. Default is 8.}

\item{warping}{If \code{nonparametric} (default), inverse warping functions are estimated nonparametrically. 
If \code{piecewise_linear2} they follow a piecewise linear function with 2 knots.}

\item{prior_1_x_mean_sd}{For \code{warping = "piecewise_linear2"} only. Optional vector of length 2 to 
specify a prior mean and sd, respectively, for a Normal distribution prior on the first knot's x location.}

\item{prior_1_y_mean_sd}{For \code{warping = "piecewise_linear2"} only. Optional vector of length 2 to 
specify a prior mean and sd, respectively, for a Normal distribution prior on the first knot's y location.}

\item{prior_2_x_mean_sd}{For \code{warping = "piecewise_linear2"} only. Optional vector of length 2 to 
specify a prior mean and sd, respectively, for a Normal distribution prior on the second knot's x location.}

\item{prior_2_y_mean_sd}{For \code{warping = "piecewise_linear2"} only. Optional vector of length 2 to 
specify a prior mean and sd, respectively, for a Normal distribution prior on the second knot's y location.}
}
\value{
The scalar value taken by the loss function.
}
\description{
Loss function for registration step optimization
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu},
Erin McDonnell \email{eim2117@cumc.columbia.edu}
}
