% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_unregistered_curves.R
\name{mean_curve}
\alias{mean_curve}
\title{Simulate mean curve}
\usage{
mean_curve(grid, period = 2 * pi, spline_based = FALSE)
}
\arguments{
\item{grid}{Grid of x values over which to evaluate the function.}

\item{period}{Controls the period of the mean curve}

\item{spline_based}{If FALSE curve is constructed using sine and cosine functions,
if TRUE, curve is constructed using B-spline basis.}
}
\value{
A numeric vector.
}
\description{
This function generates mean for simulated accelerometer data.
}
