% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R, R/registr.R
\docType{package}
\name{registr}
\alias{registr}
\title{Register Exponential Family Functional Data}
\usage{
registr(
  obj = NULL,
  Y = NULL,
  Kt = 8,
  Kh = 4,
  family = "binomial",
  gradient = TRUE,
  preserve_domain = TRUE,
  lambda_endpoint = NULL,
  beta = NULL,
  t_min = NULL,
  t_max = NULL,
  row_obj = NULL,
  periodic = FALSE,
  warping = "nonparametric",
  cores = 1L,
  ...
)
}
\arguments{
\item{obj}{Current estimate of FPC object. 
Can be NULL only if Y argument is selected.}

\item{Y}{Dataframe. Should have values id, value, index.}

\item{Kt}{Number of B-spline basis functions used to estimate mean functions. Default is 8.}

\item{Kh}{Number of B-spline basis functions used to estimate warping functions \emph{h}. Default is 4.}

\item{family}{One of \code{c("gaussian","binomial","gamma")}. Defaults to
\code{"gaussian"}.}

\item{gradient}{If \code{TRUE}, uses analytic gradient to calculate derivative. 
If \code{FALSE}, calculates gradient numerically. Not available for
\code{family = "gamma"}.}

\item{preserve_domain}{Indicator if the registration should preserve the
time domain, leading to warping functions that end on the diagonal.
Defaults to \code{TRUE}. Can only be set to \code{FALSE} when
\code{warping = "nonparametric"}.}

\item{lambda_endpoint}{Penalization parameter to control the amount of
deviation of the warping function endpoints from the diagonal. The higher
this lambda, the more the endpoints are forced towards the diagonal.
Only used if \code{preserve_domain = FALSE}.}

\item{beta}{Current estimates for beta for each subject. Default is NULL.}

\item{t_min}{Minimum value to be evaluated on the time domain.
if `NULL`, taken to be minimum observed value.}

\item{t_max}{Maximum value to be evaluated on the time domain. 
if `NULL`, taken to be maximum observed value.}

\item{row_obj}{If NULL, the function cleans the data and calculates row indices. 
Keep this NULL if you are using standalone \code{registr} function.}

\item{periodic}{If \code{TRUE}, uses periodic b-spline basis functions. Default is \code{FALSE}.}

\item{warping}{If \code{nonparametric} (default), inverse warping functions are estimated nonparametrically. 
If \code{piecewise_linear2} they follow a piecewise linear function with 2 knots.}

\item{cores}{Number of cores to be used. If \code{cores > 1}, the registration
call is parallelized by using \code{parallel::mclapply} (for Unix-based
systems) or \code{parallel::parLapply} (for Windows). Defaults to 1,
no parallelized call.}

\item{...}{additional arguments passed to or from other functions}
}
\value{
An object of class \code{fpca} containing:
\item{Y}{The observed data. The variable index is the new estimated time domain.}
\item{loss}{Value of the loss function after registraton.}
\item{beta}{Matrix of B-spline basis coefficients used to construct subject-specific warping functions.}
}
\description{
Software for registering functional data from the exponential family of distributions.

Function used in the registration step of an FPCA-based approach for 
registering exponential-family, potentially incomplete functional data,
called by \code{\link{register_fpca}}. 
This method uses constrained optimization to estimate spline 
coefficients for warping functions, where the objective function for optimization comes from 
maximizing the EF likelihood subject to monotonicity constraints on the warping functions. 
You have to either specify \code{obj}, which is a fpca 
object from an earlier step, or \code{Y}, a dataframe in long format with variables 
\code{id}, \code{index}, and \code{value} to indicate subject IDs, times, and observations, 
respectively. \cr \cr
Warping functions by default are forced to end on the diagonal to be
domain-preserving. This behavior can be changed by setting
\code{preserve_domain = FALSE} and a reasonable \code{lambda_endpoint} value.
For further details see the accompanying vignette. \cr \cr
By specifying \code{cores > 1} the registration call can be parallelized.
}
\examples{
### complete binomial curves
Y = simulate_unregistered_curves()
register_step = registr(obj = NULL, Y = Y, Kt = 6, Kh = 4, family = "binomial", 
                        gradient = TRUE)
\donttest{
### incomplete Gaussian curves
data(growth_incomplete)
library(ggplot2)

# Force the warping functions to end on the diagonal
register_step2a = registr(obj = NULL, Y = growth_incomplete, Kt = 6, Kh = 4,
                          family = "gaussian", gradient = TRUE,
                          preserve_domain = TRUE)
ggplot(register_step2a$Y, aes(x = tstar, y = index, group = id)) +
  geom_line(alpha = 0.2) +
  ggtitle("Estimated warping functions")
	
# Allow the warping functions to not end on the diagonal.
# The higher lambda_endpoint, the more the endpoints are forced towards the diagonal.
register_step2b = registr(obj = NULL, Y = growth_incomplete, Kt = 6, Kh = 4,
                          family = "gaussian", gradient = TRUE,
                          preserve_domain = FALSE, lambda_endpoint = 1)
ggplot(register_step2b$Y, aes(x = tstar, y = index, group = id)) +
  geom_line(alpha = 0.2) +
  ggtitle("Estimated warping functions")
}

}
\author{
Julia Wrobel

Julia Wrobel \email{julia.wrobel@cuanschutz.edu},
Erin McDonnell \email{eim2117@cumc.columbia.edu},
Alexander Bauer \email{alexander.bauer@stat.uni-muenchen.de}
}
