% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registr.R
\name{registr_oneCurve}
\alias{registr_oneCurve}
\title{Internal function to register one curve}
\usage{
registr_oneCurve(i, arg_list, ..., verbose = TRUE)
}
\arguments{
\item{i}{Numeric index of the curve under focus.}

\item{arg_list}{Named list of all arguments necessary for the registration
step.}

\item{...}{additional arguments passed to or from other functions}

\item{verbose}{print diagnostic messages}
}
\value{
An list containing:
\item{hinv_innerKnots}{Inner knots for setting up the spline basis
for the inverse warping function.}
\item{hinv_beta}{Estimated B-spline basis coefficients used to construct
subject-specific inverse warping functions.}
\item{t_hat}{Vector of registered time domain.}
\item{loss}{Loss of the optimal solution.}
}
\description{
This internal function is only to be used from within \code{registr}.
It performs the main optimization step with \code{constrOptim} for the
registration of one curve.
}
\author{
Julia Wrobel \email{julia.wrobel@cuanschutz.edu},
Erin McDonnell \email{eim2117@cumc.columbia.edu},
Alexander Bauer \email{alexander.bauer@stat.uni-muenchen.de}
}
