% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_register.R
\name{rr_register}
\alias{rr_register}
\title{Download a register}
\usage{
rr_register(file, phase = c("beta", "alpha"), path_type = c("url", "file"),
  parse_datetimes = FALSE)
}
\arguments{
\item{file}{Character, name of the register, e.g. "school-eng"}

\item{phase}{Character, one of "beta", "alpha", default: "beta"}

\item{path_type}{Character, one of \code{"url"} or \code{"file"} to decide what to do
with \code{register}.}

\item{parse_datetimes}{Logical, whether to parse ISO8601 strings as datetimes
with \code{\link[parsedate:parse_iso_8601]{parsedate::parse_iso_8601()}}, otherwise leave as a string.  Partial
datetimes are parsed as the earliest possible datetime, e.g. \code{"2018"}
becomes \code{"2018-01-01 UTC"}.}
}
\value{
An S3 object of class \code{register}
}
\description{
Downloads a whole register (or reads it from a file) and
constructs an object that can be interrogated for its records, entries,
items, schema, links to other registers, etc.

You should probably run \code{\link[=rr_snapshot]{rr_snapshot()}} on the output before using it.
}
\examples{
\dontrun{
  rr_register("country")
  rr_register("country", "beta")
  path <- tempfile()
  download.file("https://country.register.gov.uk/download-rsf", path)
  rr_register(path, path_type = "file")
  unlink(path)
}


}
