% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_register.R
\name{rr_register}
\alias{rr_register}
\title{Download a register}
\usage{
rr_register(register, phase = "beta")
}
\arguments{
\item{register}{character, name of the register, e.g. "school-eng"}

\item{phase}{character, one of "beta", "alpha", default: "beta"}
}
\value{
An S3 object of class \code{register}
}
\description{
Downloads a whole register and constructs an object that can be
interrogated for its records, entries, items, schema, links to other
registers, etc.

You should probably run \code{\link[=rr_snapshot]{rr_snapshot()}} on the output before using it.
}
\examples{
\dontrun{
  rr_register("country")
  rr_register("country", "beta")
}

}
