% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_registers.R
\name{rr_registers}
\alias{rr_registers}
\title{Download all registers from an environment}
\usage{
rr_registers(phase = c("beta", "alpha"), dir = NULL,
  parse_datetimes = FALSE, write = FALSE, dest_dir = phase)
}
\arguments{
\item{phase}{Character, one of \code{"beta"} or \code{"alpha"}}

\item{dir}{Character, optional name of a directory of RSF files of registers.
If \code{NULL} (default), the registers will be downloaded.}

\item{parse_datetimes}{Logical, whether to parse ISO8601 strings as datetimes
with \code{\link[parsedate:parse_iso_8601]{parsedate::parse_iso_8601()}}, otherwise leave as a string.  Partial
datetimes are parsed as the earliest possible datetime, e.g. \code{"2018"}
becomes \code{"2018-01-01 UTC"}.}

\item{write}{Logical, whether to write the RSF file to disk.  If \code{TRUE},
either \code{name} or \code{dest_path} must be provided.}

\item{dest_dir}{Character, name of a directory to save the RSF file of each
register to.  The file names will be the names of the registers.}
}
\value{
An S3 object of class \code{register}
}
\description{
Downloads all the registers listed in the register register of
an environment ('alpha' or 'beta', meaning 'ready to use' or 'open for
feedback').

Unlike \code{\link[=rr_register]{rr_register()}} this function only downloads register, and does not
read them from disk.
}
\examples{
path <- tempdir()
names(rr_registers("beta", write = TRUE, dest_dir = path))
names(rr_registers(dir = path))
}
